Edit Role & Permission Section

+++ app/Http/Controllers/ProductController.php
@@ -28,7 +28,16 @@ class ProductController extends Controller
     
     public function store(StoreProductRequest $request)
     {
-        $porduct = Product::create($request->toArray());
+        $addedBy = auth()->user()->id;
+    
+        $porduct = Product::create([
+            'productName'=>$request->productName,
+            'number'=>$request->number,
+            'price'=>$request->price,
+            'color'=>$request->color,
+            'addedBy'=>$addedBy,
+            'description'=>$request->description,
+        ]);
 
         $categories = $request->categories;
 
@@ -71,51 +80,58 @@ class ProductController extends Controller
     
     public function update(UpdateProductRequest $request, string $id)
     {
+        $userLoggedIn = auth()->user();
         $porduct = Product::find($id);
-
-        if (!$porduct) {
-           return response()->json([
-            'message' => 'porduct not found',
-            'status'=> 'Not Found'
-           ] , 404);
-        }
-
-        $porduct->update($request->toArray());
-
-        if($request->categories)
-        {
-            $categories = $request->categories;
-            $categorySync = [];
-            foreach($categories as $category)
-            {
-               $categorySync[] = $category;
-            }
-            $porduct->categories()->attach($categorySync);
-        }
         
-        return response()->json([
-            'porduct'=>$porduct,
-            'status'=> 'success'
-         ] , 200);
-    }
+        if(!$porduct){
+            return response()->json([
+                'message' => 'porduct not found',
+                'status'=> 'Not Found'
+            ] , 404); 
+        }
 
+        elseif($userLoggedIn->hasPermissionTo('products.update') || ($userLoggedIn->hasPermissionTo('products.update.seler') && $userLoggedIn->id == $porduct->addedBy)){
+            
+            $porduct->update($request->toArray());
+
+            if($request->categories)
+            {
+                $categories = $request->categories;
+                $categorySync = [];
+                foreach($categories as $category)
+                {
+                $categorySync[] = $category;
+                }
+                $porduct->categories()->attach($categorySync);
+            }
+            
+            return response()->json([
+                'porduct'=>$porduct,
+                'status'=> 'success'
+            ] , 200);
+        }
+   }
 
 
     public function destroy(string $id)
     {
+        $userLoggedIn = auth()->user();
         $porduct = Product::find($id);
-
-        if (!$porduct) {
+        
+        if(!$porduct){
             return response()->json([
-             'message' => 'porduct not found',
-             'status'=> 'Not Found'
-            ] , 404);
-         }
+                'message' => 'porduct not found',
+                'status'=> 'Not Found'
+            ] , 404); 
+        }
+
+        elseif($userLoggedIn->hasPermissionTo('products.update') || ($userLoggedIn->hasPermissionTo('products.update.seler') && $userLoggedIn->id == $porduct->addedBy)){
 
          $porduct->delete();
          return response()->json([
             'message' => 'porduct destroy successfully',
             'status' => 'success'
          ] , 200);
+        }
     }
 }
